balancing

+++ src/beings/Humanoid.java
@@ -19,7 +19,7 @@ public abstract class Humanoid implements Drawable, Collidable, Destructable, Up
     protected float maxSpeed;
     protected boolean alive;
 
-    public Humanoid(int x, int y, int maxSpeed) {
+    public Humanoid(int x, int y, float maxSpeed) {
         this.health = 100;
         this.maxHealth = 100;
         alive = true;
